NN-3467: - move restricted patient service and related logic into PrisonerIndexService.kt
