+++ src/main/kotlin/uk/gov/justice/digital/hmpps/prisonersearch/services/RestrictedPatientService.kt
@@ -1,9 +1,9 @@
 package uk.gov.justice.digital.hmpps.prisonersearch.services
 
-import io.netty.handler.codec.http.HttpResponseStatus.NOT_FOUND
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
+import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
 import org.springframework.web.reactive.function.client.WebClient
 import org.springframework.web.reactive.function.client.WebClientResponseException
@@ -12,14 +12,17 @@ import uk.gov.justice.digital.hmpps.prisonersearch.services.dto.RestrictedPatien
 @Service
 @ConditionalOnProperty(value = ["api.base.url.restricted-patients"])
 class RestrictedPatientService(@Qualifier("restrictedPatientsWebClient") val webClient: WebClient) {
-  fun getRestrictedPatient(prisonerNumber: String): RestrictedPatientDto? = try {
-    webClient.get().uri("/restricted-patient/prison-number/$prisonerNumber")
-      .retrieve()
-      .bodyToMono(RestrictedPatientDto::class.java)
-      .block()
-  } catch (e: WebClientResponseException) {
-    if (e.statusCode.equals(NOT_FOUND)) null
-    throw e
+  fun getRestrictedPatient(prisonerNumber: String): RestrictedPatientDto? {
+    try {
+      return webClient.get().uri("/restricted-patient/prison-number/$prisonerNumber")
+        .retrieve()
+        .bodyToMono(RestrictedPatientDto::class.java)
+        .block()
+    } catch (e: WebClientResponseException) {
+      if (e.statusCode.equals(HttpStatus.NOT_FOUND)) return null
+
+      throw e
+    }
   }
 }
